[Feat:#219] - 리스트,필터 390px기준으로 변경
